gui/requirements.txt
matplotlib==3.6.2
pandas==1.5.2
Pillow==9.5.0
pyautogui==0.9.53
pyshark==0.5.3
